% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_methods.R
\name{bridge-methods}
\alias{bridge-methods}
\alias{summary.bridge}
\alias{summary.bridge_list}
\alias{print.summary.bridge}
\alias{print.summary.bridge_list}
\alias{print.bridge}
\alias{print.bridge_list}
\title{Methods for bridge and bridge_list objects}
\usage{
\method{summary}{bridge}(object, na.rm = TRUE, ...)

\method{summary}{bridge_list}(object, na.rm = TRUE, ...)

\method{print}{summary.bridge}(x, ...)

\method{print}{summary.bridge_list}(x, ...)

\method{print}{bridge}(x, ...)

\method{print}{bridge_list}(x, na.rm = TRUE, ...)
}
\arguments{
\item{object, x}{object of class \code{bridge} or \code{bridge_list} as returned from \code{\link{bridge_sampler}}.}

\item{na.rm}{logical. Should NA estimates in \code{bridge_list} objects be removed? Passed to \code{\link{error_measures}}.}

\item{...}{further arguments, currently ignored.}
}
\value{
The \code{summary} methods return a \code{data.frame} which contains the log marginal likelihood plus the result returned from invoking \code{\link{error_measures}}.

The \code{print} methods simply print and return nothing.
}
\description{
Methods defined for objects returned from the generic \code{\link{bridge_sampler}} function.
}
