% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.sampler.R
\name{bridge.sampler}
\alias{bridge.sampler}
\title{Computes log marginal likelihood via bridge sampling}
\usage{
bridge.sampler(post.samples = NULL, log.posterior = NULL, data = NULL,
  lb = NULL, ub = NULL, method = "normal", cores = 1, packages = NULL,
  rcppFile = NULL)
}
\arguments{
\item{post.samples}{matrix with posterior samples (colnames need to correspond to parameter names)}

\item{log.posterior}{function for computing log of prior times likelihood}

\item{data}{data}

\item{lb}{named vector with lower bounds for parameters}

\item{ub}{named vector with upper bounds for parameters}

\item{method}{either "normal" or "warp3"}

\item{cores}{number of cores used for evaluating log.posterior}

\item{packages}{character vector with names of packages needed for evaluating log.posterior in parallel (only relevant if cores > 1)}

\item{rcppFile}{in case cores > 1 and log.posterior is an Rcpp function, rcppFile specifies the path to the cpp file (needs to be compiled on all cores)}
}
\value{
a list of the class \code{bridge} with the objects:
\itemize{
 \item \code{logml}: estimate of log marginal likelihood
 \item \code{niter}: number of iterations of the iterative updating scheme
 \item \code{method}: bridge sampling method that was used to obtain estimate
 \item \code{q11}: log.posterior evaluations for posterior samples
 \item \code{q12}: log proposal evaluations for posterior samples
 \item \code{q21}: log.posterior evaluations for samples from proposal
 \item \code{q22}: log proposal evaluations for samples from proposal
}
}
\description{
Computes log marginal likelihood via bridge sampling
}
\details{
Bridge sampling is implemented as described in Meng and Wong (1996, see equation 4.1) using the "optimal" bridge function. When \code{method = "normal"}, the proposal distribution is a multivariate normal distribution with mean vector equal to the column means of \code{post.samples} and covariance matrix equal to the sample covariance matrix of \code{post.samples}. When \code{method = "warp3"},
the proposal distribution is a standard multivariate normal distribution and the posterior distribution is "warped" (Meng & Schilling, 2002) so that it has the same mean vector, covariance matrix and skew as post.samples. \code{method = "warp3"} usually yields more precise results, but it takes approximately twice as long as \code{method = "normal"}.
}
\references{
Meng, X.-L., & Wong, W. H. (1996). Simulating ratios of normalizing constants via a simple identity: A theoretical exploration. Statistica Sinica, 6, 831-860.

Meng, X.-L., & Schilling, S. (2002). Warp bridge sampling. Journal of Computational and Graphical Statistics, 11(3), 552-586.
}
\author{
Quentin F. Gronau
}
