% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bf.R
\name{compute_bf}
\alias{compute_bf}
\title{Computing the Bayes factor}
\usage{
compute_bf(bridgeObject1, bridgeObject2, log = FALSE)
}
\arguments{
\item{bridgeObject1}{output from \code{\link{bridge_sampler}}.}

\item{bridgeObject2}{output from \code{\link{bridge_sampler}}.}

\item{log}{Boolean. If \code{TRUE}, the function returns the log of the Bayes factor. Default is \code{FALSE}.}
}
\value{
Value of the Bayes factor (scalar).
}
\description{
Computes the Bayes factor for a model compairson between two models.
}
\details{
Computes the Bayes factor (Kass & Raftery, 1995) in favor of the model associated with \code{bridgeObject1} over the model associated with \code{bridgeObject2}.
}
\note{
For examples, see \code{\link{bridge_sampler}} and the accompanying vignette: \code{vignette("bridgesampling_example")}.
}
\references{
Kass, R. E., & Raftery, A. E. (1995). Bayes factors. Journal of the american statistical association, 90(430), 773-795.
}
\author{
Quentin F. Gronau
}
