% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_post_prob.R
\name{compute_post_prob}
\alias{compute_post_prob}
\title{Computing posterior model probabilities}
\usage{
compute_post_prob(..., prior_prob)
}
\arguments{
\item{...}{outputs from \code{\link{bridge_sampler}} of class \code{bridge}.}

\item{prior_prob}{numeric vector with prior model probabilities. If omitted a uniform prior is used (i.e., all models equally likely a priori).}
}
\value{
numeric vector with posterior model probabilities.
}
\description{
Computes posterior model probabilities.
}
\author{
Quentin F. Gronau
}
