% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{error_measures}
\alias{error_measures}
\title{Computing error measures for estimated marginal likelihood}
\usage{
error_measures(bridgeObject)
}
\arguments{
\item{bridgeObject}{output from \code{\link{bridge_sampler}}.}
}
\value{
a list with the objects:
\itemize{
 \item \code{re2}: approximate relative mean squared error for marginal likelihood estimate
 \item \code{cv}: coefficient of variation for marginal likelihood estimate (assumes that bridge estimate is unbiased)
 \item \code{percentage}: percentage error of marginal likelihood estimate
}
}
\description{
Computes error measures for estimated marginal likelihood.
}
\details{
Computes approximate error measures for marginal likelihood bridge sampling estimates. Based on Fruehwirth-Schnatter (2004).
}
\note{
For examples, see \code{\link{bridge_sampler}} and the accompanying vignette.
}
\references{
Frühwirth‐Schnatter, S. (2004). Estimating marginal likelihoods for mixture and Markov switching models using bridge sampling techniques. The Econometrics Journal, 7, 143-167.
}
\author{
Quentin F. Gronau
}
