% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logml.R
\name{logml}
\alias{logml}
\alias{logml.bridge}
\alias{logml.bridge_list}
\title{Log Marginal Likelihoods from Bridge Objects}
\usage{
logml(x, ...)

\method{logml}{bridge}(x, ...)

\method{logml}{bridge_list}(x, fun = median, ...)
}
\arguments{
\item{x}{Object of class \code{"bridge"} or \code{"bridge_list"} as returned from \code{\link{bridge_sampler}}.}

\item{...}{Further arguments passed to \code{fun}.}

\item{fun}{Function which returns a scalar value and is applied to the \code{logml} vector of \code{"bridge_list"} objects. Default is \code{\link{median}}.}
}
\value{
scalar numeric
}
\description{
Generic function that returns log marginal likelihood from bridge objects. For objects of class \code{"bridge_list"}, which contains multiple log marginal likelihoods, \code{fun} is performed on the vector and its result returned.
}
