% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_post_prob.R
\name{post_prob}
\alias{post_prob}
\alias{post_prob.bridge}
\alias{post_prob.default}
\title{Posterior Model Probabilities from Marginal Likelihoods}
\usage{
post_prob(x, ..., prior_prob)

\method{post_prob}{bridge}(x, ..., prior_prob)

\method{post_prob}{default}(x, ..., prior_prob)
}
\arguments{
\item{x}{Object of class \code{"bridge"} or \code{"bridge_list"} as returned from \code{\link{bridge_sampler}}. The default method simply assumes all passed objects are numeric log marginal likelihoods (e.g., from \code{\link{logml}}).}

\item{...}{further objects of class \code{"bridge"} or \code{"bridge_list"} as returned from \code{\link{bridge_sampler}}. Or numeric values for the default method.}

\item{prior_prob}{numeric vector with prior model probabilities. If omitted, a uniform prior is used (i.e., all models are equally likely a priori).}
}
\value{
Numeric vector with posterior model probabilities (with names derived from the input).
}
\description{
Generic function that computes posterior model probabilities from marginal likelihoods
}
\note{
For realistic examples, see \code{\link{bridge_sampler}} and the accompanying vignette: \cr \code{vignette("bridgesampling_example")}
}
\examples{

H0 <- structure(list(logml = -20.8084543022433, niter = 4, method = "normal"),
                .Names = c("logml", "niter", "method"), class = "bridge")
H1 <- structure(list(logml = -17.9623077558729, niter = 4, method = "normal"),
                .Names = c("logml", "niter", "method"), class = "bridge")
H2 <- structure(list(logml = -19, niter = 4, method = "normal"),
                .Names = c("logml", "niter", "method"), class = "bridge")


post_prob(H0, H1, H2)
post_prob(H1, H0)

## all produce the same (only names differ):
post_prob(H0, H1, H2)
post_prob(H0$logml, H1$logml, H2$logml)
post_prob(c(H0$logml, H1$logml, H2$logml))
post_prob(H0$logml, c(H1$logml, H2$logml))


}
\author{
Quentin F. Gronau and Henrik Singmann
}
