% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_bridge_sampler.R
\name{stan_bridge_sampler}
\alias{stan_bridge_sampler}
\title{Marginal likelihood for an object of class \code{stanfit}}
\usage{
stan_bridge_sampler(stanfit_data = NULL, stanfit_model = stanfit_data,
  repetitions = 1, method = "normal", cores = 1, maxiter = 1000,
  silent = FALSE, verbose = FALSE)
}
\arguments{
\item{stanfit_data}{an object of class \code{"stanfit"} that contains the posterior samples.}

\item{stanfit_model}{an object of class \code{"stanfit"} with the same model as \code{stanfit_data}, which will be used for evaluating the \code{log_posterior} (i.e., it does not need to contain any samples). The default is to use \code{stanfit_data}. In case \code{stanfit_data} was compiled on another computer with a different OS or setup, \code{stanfit_data} usually cannot be used for evaluation. In this case, one can compile the model on the current computer with \code{iter = 0} and pass it here (this usually needs to be done before \code{stanfit_data} is loaded).}

\item{repetitions}{number of repetitions.}

\item{method}{either \code{"normal"} or \code{"warp3"}.}

\item{cores}{number of cores used for computations. \code{cores > 1} is currently only supported on unix-like systems that support forking via \code{\link{mclapply}} (e.g., Linux or Mac OS).}

\item{maxiter}{maximum number of iterations for the iterative updating scheme. Default is 1,000 to avoid infinite loops.}

\item{silent}{Boolean which determines whether to print the number of iterations of the updating scheme to the console. Default is FALSE.}

\item{verbose}{Boolean. Should internal debug information be printed to console? Default is FALSE.}
}
\value{
a list of class \code{"bridge"} with components:
\itemize{
 \item \code{logml}: estimate of log marginal likelihood.
 \item \code{niter}: number of iterations of the iterative updating scheme.
 \item \code{method}: bridge sampling method that was used to obtain the estimate.
 \item \code{q11}: log_posterior evaluations for posterior samples.
 \item \code{q12}: log proposal evaluations for posterior samples.
 \item \code{q21}: log_posterior evaluations for samples from proposal.
 \item \code{q22}: log proposal evaluations for samples from proposal.
}
}
\description{
Computes marginal likelihood for an object of class \code{stanfit}.
}
\details{
Works.
}
\author{
Quentin F. Gronau and Henrik Singmann. Uses code from \code{rstan} by Jiaqing Guo, Jonah Gabry, and Ben Goodrich.
}
