% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_post_prob.R
\name{compute_post_prob}
\alias{compute_post_prob}
\title{Posterior Model Probabilities from Marginal Likelihoods}
\usage{
compute_post_prob(..., prior_prob)
}
\arguments{
\item{...}{objects of class \code{"bridge"} as returned from \code{\link{bridge_sampler}}.}

\item{prior_prob}{numeric vector with prior model probabilities. If omitted, a uniform prior is used (i.e., all models are equally likely a priori).}
}
\value{
Numeric vector with posterior model probabilities.
}
\description{
Computes posterior model probabilities.
}
\note{
For examples, see \code{\link{bridge_sampler}} and the accompanying vignette: \cr \code{vignette("bridgesampling_example")}
}
\author{
Quentin F. Gronau
}
