% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_bridge_sampler.R
\name{stan_bridge_sampler}
\alias{stan_bridge_sampler}
\title{Marginal likelihood for an object of class \code{stanfit}}
\usage{
stan_bridge_sampler(stanfit = NULL, method = "normal", cores = 1,
  maxiter = 1000, silent = FALSE)
}
\arguments{
\item{stanfit}{an object of class \code{"stanfit"}.}

\item{method}{either \code{"normal"} or \code{"warp3"}.}

\item{cores}{number of cores used for computations.}

\item{maxiter}{maximum number of iterations for the iterative updating scheme. Default is 1,000 to avoid infinite loops.}

\item{silent}{Boolean which determines whether to print the number of iterations of the updating scheme to the console. Default is FALSE.}
}
\value{
a list of class \code{"bridge"} with components:
\itemize{
 \item \code{logml}: estimate of log marginal likelihood.
 \item \code{niter}: number of iterations of the iterative updating scheme.
 \item \code{method}: bridge sampling method that was used to obtain the estimate.
 \item \code{q11}: log_posterior evaluations for posterior samples.
 \item \code{q12}: log proposal evaluations for posterior samples.
 \item \code{q21}: log_posterior evaluations for samples from proposal.
 \item \code{q22}: log proposal evaluations for samples from proposal.
}
}
\description{
Computes marginal likelihood for an object of class \code{stanfit}.
}
\details{
Computes approximate error measures for marginal likelihood bridge sampling estimates. Based on Fruehwirth-Schnatter (2004).
}
\author{
Quentin F. Gronau
}
